+++ main.py
@@ -434,9 +434,6 @@ class OutlierRemoval:
         self.plot_data()
 
 
-
-
-
 if __name__ == '__main__':
     root = tk.Tk()
     window = TrackingUI(root)